moves scan functions to scan_task module
