+++ zebra-scan/src/lib.rs
@@ -9,12 +9,14 @@ extern crate tracing;
 
 pub mod config;
 pub mod init;
-pub mod scan;
 pub mod storage;
 
 use zebra_node_services::scan_service::{request::Request, response::Response};
 
 pub mod service;
+
+pub use service::scan_task::scan;
+
 #[cfg(any(test, feature = "proptest-impl"))]
 pub mod tests;
